+++ tcod/ec/__init__.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 __version__ = "1.2.0"
 
 import reprlib
+import warnings
 from typing import Any, Dict, Iterable, Iterator, Optional, Type, TypeVar
 
 from typing_extensions import Self
@@ -20,6 +21,25 @@ def abstract_component(cls: Type[T]) -> Type[T]:
 
     Subclasses of this `cls` will now use `cls` as the key when being accessed in :any:`ComponentDict`.
     This means that ComponentDict can only hold one unique instance of this subclass.
+
+    .. deprecated:: Unreleased
+
+        This method of handling abstract components was deemed unnecessary.
+        Instead a new component should be made to hold the base class, for example::
+
+            >>> from tcod.ec import ComponentDict
+            >>> import attrs
+            >>> @attrs.define
+            ... class Base:
+            ...     pass
+            >>> @attrs.define
+            ... class Derived(Base):
+            ...     pass
+            >>> @attrs.define
+            ... class Abstract(Base):
+            ...     value: Base
+            >>> ComponentDict([Abstract(Derived())])
+            ComponentDict([Abstract(value=Derived())])
 
     Example::
 
@@ -48,6 +68,12 @@ def abstract_component(cls: Type[T]) -> Type[T]:
         >>> entity[Base]
         Derived()
     """
+    warnings.warn(
+        "The abstract_component decorator is deprecated."
+        "  To store a derived class: make a new component which stores the base class and holds the derived object.",
+        FutureWarning,
+        stacklevel=2,
+    )
     cls._COMPONENT_TYPE = cls  # type: ignore[attr-defined]
     return cls